+++ src/app/app-routing.module.ts
@@ -3,14 +3,12 @@ import { RouterModule, Routes } from '@angular/router';
 
 // Components
 import { NetworkComponent } from './network/network.component';
-import { WifiComponent } from './wifi/wifi.component';
 import { CommandsComponent } from './commands/commands.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 
 const routes: Routes = [
   { path: "", redirectTo: "/network", pathMatch: "full" },
   { path: "network", component: NetworkComponent },
-  { path: "wifi", component: WifiComponent },
   { path: "commands", component: CommandsComponent },
   { path: '**', component: PageNotFoundComponent },
 ];