+++ src/app/interfaces/StationWifiIntf.ts
@@ -1,7 +1,0 @@
-export interface StationWifiIntf {
-  "enable_wifi": string,
-  "network_name": string,
-  "network_password": string,
-  "enable_bridge": string,
-  "network_ap_gateway": string,
-}